The abstract needs to be rewritten
I have inserted some comments, but the abstract is all wrong, including things that shouldn't go in the abstract (GP is this... GP works this way) and saying nothing on what is the problam you want to solve and why you want to solve it this way.

Please @fergunet @amorag @pacastillo rewrite the abstract along these lines.

This is the problem we have (evolution needs to search a big search space, not restraining it to a series of coefficients, for instance). GP evolves bots in free form, and...
But evolution needs guidance in the shape of a fitness function. Although a priori using direct score or number of victories might seem the best option, it is not because...

That is why AND THIS IS OUR MAIN SELLING POINT in this paper we have evaluated different fitness functions that use this or that. We have tested them on this game because it represents...

Our results show that...

--

And not
We are using GP. GP is cool. We are applying it to this game and testing (deeply) 3 fitness. These are the results.

Please!