Added some suggestions
But basically I don't understand what the new graphs are about. If you are talking about the behavior of GP, that is an entirely new paper and if it goes in this paper it should go _before_ you talk about the fitnesses. If you are researching on what GP evolves, that is an interesting paper, but a _different_ paper. So please @deantares @fergunet try to explain this.