English, please
"reinforce troops" does not exist.
When I say "what is reinforces" it is not because I don't understand English or the underlying concept. It is because it is not a thing in English. reinforce is a verb, not a subject. Please @deantares or @amorag use the right word for this. It is probably "reinforcements", but I have only found that in the context of Clash of Clans. Please @amorag if you do not know a word do not invent it and make witty comments to whoever asks you what you mean, look it up in the dictionary.