Please understand the state of the art and conclusions @amorag @deantares @fergunet
You already proved in 2014 that GP could be used for Planet Wars. The first sentence of the conclusions:
```
The objective of this work is to validate if using Genetic Programming
can create competitive bots for RTS game
```
is simply not valid. This paper must be about something else, and I insist: please a single issue per paper. This paper is either about fitness or about insights from the evolved trees. Not about both. You can't deal shallowly with two topics in a single paper. You should study a single topic in depth.